+++ test/e2e/selective_sync_test.go
@@ -8,12 +8,12 @@ import (
 	"github.com/argoproj/gitops-engine/pkg/health"
 	. "github.com/argoproj/gitops-engine/pkg/sync/common"
 
-	. "github.com/argoproj/argo-cd/pkg/apis/application/v1alpha1"
-	"github.com/argoproj/argo-cd/test/e2e/fixture"
-	. "github.com/argoproj/argo-cd/test/e2e/fixture"
-	. "github.com/argoproj/argo-cd/test/e2e/fixture/app"
-	. "github.com/argoproj/argo-cd/util/errors"
-	"github.com/argoproj/argo-cd/util/rand"
+	. "github.com/argoproj/argo-cd/v2/pkg/apis/application/v1alpha1"
+	"github.com/argoproj/argo-cd/v2/test/e2e/fixture"
+	. "github.com/argoproj/argo-cd/v2/test/e2e/fixture"
+	. "github.com/argoproj/argo-cd/v2/test/e2e/fixture/app"
+	. "github.com/argoproj/argo-cd/v2/util/errors"
+	"github.com/argoproj/argo-cd/v2/util/rand"
 )
 
 // when you selectively sync, only selected resources should be synced, but the app will be out of sync
@@ -53,7 +53,9 @@ func TestSelectiveSyncDoesNotRunHooks(t *testing.T) {
 func TestSelectiveSyncWithoutNamespace(t *testing.T) {
 	selectedResourceNamespace := getNewNamespace(t)
 	defer func() {
-		FailOnErr(Run("", "kubectl", "delete", "namespace", selectedResourceNamespace))
+		if !t.Skipped() {
+			FailOnErr(Run("", "kubectl", "delete", "namespace", selectedResourceNamespace))
+		}
 	}()
 	Given(t).
 		Prune(true).
@@ -81,7 +83,9 @@ func TestSelectiveSyncWithoutNamespace(t *testing.T) {
 func TestSelectiveSyncWithNamespace(t *testing.T) {
 	selectedResourceNamespace := getNewNamespace(t)
 	defer func() {
-		FailOnErr(Run("", "kubectl", "delete", "namespace", selectedResourceNamespace))
+		if !t.Skipped() {
+			FailOnErr(Run("", "kubectl", "delete", "namespace", selectedResourceNamespace))
+		}
 	}()
 	Given(t).
 		Prune(true).
